requirements.txt
alembic==1.11.1
aniso8601==9.0.1
atomicwrites==1.4.1
attrs==23.1.0
blinker==1.6.2
click==8.1.3
colorama==0.4.6
Flask==2.3.2
Flask-JWT-Extended==4.5.2     
Flask-Migrate==4.0.4
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.0.5       
greenlet==2.0.2
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
mysql-connector-python==8.0.33
packaging==23.1
passlib==1.7.4
pluggy==1.2.0
protobuf==3.20.3
psycopg2-binary==2.9.6        
py==1.11.0
PyJWT==2.7.0
pytest==7.1.2
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
SQLAlchemy==2.0.17
tomli==2.0.1
typing_extensions==4.7.1      
uuid==1.30
Werkzeug==2.3.6